RISE/Backend Server/requirements.txt
python-dotenv
flask
langchain
langchain-community
openai
pinecone-client[grpc]
langchain-openai==0.2.3
langchain-pinecone==0.2.0